+++ backends/carp_webservices/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.11.1
-* support for authentication using a form (UI).
 * the singleton CarpService is accessed using `CarpService()` (instead of the old `CarpService.instance`) 
   in order to be consist with the rest of CAMS.
+* support for authentication using a form (UI).
+* support for sending 'forgotten password' emails using the `sendForgottenPasswordEmail()` method.
 
 ## 0.11.0
 * updated to carp_mobile_sensing v. 0.11.x